requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
asyncpg==0.29.0
sqlalchemy==2.0.30
databases==0.9.0
pydantic==2.7.1
python-dotenv==1.0.1
